+++ src/main/java/de/dealog/msg/messaging/tracking/MessageTrackingType.java
@@ -1,0 +1,17 @@
+package de.dealog.msg.messaging.tracking;
+
+/**
+ * The available message tracking types
+ */
+public enum MessageTrackingType {
+
+    /**
+     * A single message is requested
+     */
+    SingleMessageRequested,
+
+    /**
+     * A list of messages are requested
+     */
+    ListRequested;
+}