+++ src/main/java/de/dealog/msg/rest/validations/ValidGeoRequest.java
@@ -1,0 +1,20 @@
+package de.dealog.msg.rest.validations;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.*;
+
+@Constraint(validatedBy = ValidGeoRequestValidator.class)
+@Target({ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface ValidGeoRequest {
+
+    String MESSAGE = "To query by location both parameter 'long' and 'lat'  must be provided";
+
+    String message() default MESSAGE;
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}