+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 1.0.0 (under development)
 This is the initial version setting up the base system
 
-+- 'CNG' Prefix content type of REST resources with 'vnd.'
+- 'CNG' Prefix content types of REST resources with 'vnd.'
+- `NEW` Add single and list message request tracking
+- `NEW` Add REST parameter validations
+- `NEW` Add max page size to 50
+- `NEW` Add 0.0.0 as unsupported version
 - `NEW` Add region code to message and ability to find messages by region code
 - `NEW` Add region resource to find regions by region code or location
 - `NEW` Add REST method to return a region hierachy by ars or location